Jcollard/add doc gen (#11)
Created a utility for generating documentation and deploying that documentation to an S3 bucket.

The documentation tool utilizes pdoc3 which generates HTML (or PDF) documentation files for a python module. The documentation style we will use moving forward is the Google Python Style.

To run the utility:

$ bash utils/doc.sh
Used for generating documentation for the project

Usage:
  utils/doc.sh generate - Runs pydoc over all modules
  utils/doc.sh deploy - Using aws-cli, deploys documentation to an S3 bucket

Environment Variables:
  BUCKET_NAME - The name of the bucket to be deployed to (Default: current branch name)
Files are placed in html/{module_name}. If a new module is created, it should be added to two places: html/index.html and utils/doc.sh (similar to lint).